+++ dune/typetree/powernode.hh
@@ -62,10 +62,6 @@ namespace Dune {
       //! Mark this class as a non composite in the \ref TypeTree.
       static const bool isComposite = false;
 
-      //! The number of children.
-      [[deprecated("Will be removed after release 2.9. Use degree()")]]
-      static const std::size_t CHILDREN = k;
-
       static constexpr auto degree ()
       {
         return std::integral_constant<std::size_t,k>{};